10.18 html,css구현,likelist localhost접속성공, 단위테스트 성공/ 기능구현 순서대로 다시 하기

+++ src/main/java/com/example/anysale/likeList/service/LikeListService.java
@@ -13,5 +13,5 @@ public interface LikeListService {
 
   void removeLikeList(String id);
 
-  LikeList removeLikeList(Long likeListId);
+  LikeList removeLikeList(int likeListId);
 }
